requirements.txt
gin-config~=0.5.0
sacrebleu~=2.5.1
psutil~=6.1.1
requests~=2.32.3
numpy~=2.2.2
six~=1.17.0
snowballstemmer~=3.0.1
nltk~=3.9.1
scipy~=1.15.1
trafilatura~=2.0.0
beautifulsoup4~=4.13.3
selenium~=4.34.2
jupyter~=1.1.1
matplotlib~=3.10.0
pandas~=2.2.3
seaborn~=0.13.2
stanza~=1.10.1
krippendorff~=0.8.1
scikit-learn~=1.6.1
datasets~=4.0.0
networkx~=3.4.2
tqdm~=4.67.1
bert-score~=0.3.13
transformers~=4.54.1
spacy~=3.8.7
vllm~=0.10.0
prometheus_eval
git+https://github.com/bheinzerling/pyrouge.git